Refactor service api package
