+++ transaction-service-api/src/main/java/pl/adamboguszewski/transaction/service/api/transaction/CreateTransactionRequest.java
@@ -1,8 +1,8 @@
-package pl.adamboguszewski.transaction.service.service.api.transaction;
+package pl.adamboguszewski.transaction.service.api.transaction;
 
 import lombok.Value;
-import pl.adamboguszewski.transaction.service.service.api.Currency;
-import pl.adamboguszewski.transaction.service.service.api.PaymentType;
+import pl.adamboguszewski.transaction.service.api.Currency;
+import pl.adamboguszewski.transaction.service.api.PaymentType;
 
 import javax.validation.constraints.NotNull;
 import java.time.LocalDateTime;
