--- transaction-service-api/src/main/java/pl/adamboguszewski/transaction/service/service/api/transaction/CreateTransactionResponse.java
@@ -1,4 +1,0 @@
-package pl.adamboguszewski.transaction.service.service.api.transaction;
-
-public interface CreateTransactionResponse {
-}
