+++ transaction-service/src/main/java/pl/adamboguszewski/transaction/service/application/PaymentInformation.java
@@ -1,9 +1,9 @@
 package pl.adamboguszewski.transaction.service.application;
 
 import lombok.Value;
-import pl.adamboguszewski.transaction.service.service.api.Currency;
-import pl.adamboguszewski.transaction.service.service.api.PaymentType;
-import pl.adamboguszewski.transaction.service.service.api.transaction.CreateTransactionRequest;
+import pl.adamboguszewski.transaction.service.api.Currency;
+import pl.adamboguszewski.transaction.service.api.PaymentType;
+import pl.adamboguszewski.transaction.service.api.transaction.CreateTransactionRequest;
 
 @Value
 public class PaymentInformation {
