+++ transaction-service/src/main/java/pl/adamboguszewski/transaction/service/application/Transaction.java
@@ -1,7 +1,7 @@
 package pl.adamboguszewski.transaction.service.application;
 
 import lombok.Value;
-import pl.adamboguszewski.transaction.service.service.api.transaction.CreateTransactionRequest;
+import pl.adamboguszewski.transaction.service.api.transaction.CreateTransactionRequest;
 
 import java.util.List;
 import java.util.UUID;
