+++ transaction-service/src/main/java/pl/adamboguszewski/transaction/service/application/TransactionInformation.java
@@ -1,7 +1,7 @@
 package pl.adamboguszewski.transaction.service.application;
 
 import lombok.Value;
-import pl.adamboguszewski.transaction.service.service.api.transaction.CreateTransactionRequest;
+import pl.adamboguszewski.transaction.service.api.transaction.CreateTransactionRequest;
 
 import java.time.LocalDateTime;
 import java.util.List;
